pixelstar: Drop our custom soong header generator
Enough significant changes have been made to soong in Android 12
for us to continue using this without a potential rewired logic.
On top of that, we've decided internally to use device_kernel_headers
and qti_kernel_headers as they are gonna replace our generated_kernel_headers.
Good bye this goes.

Change-Id: I3029579ec63fc55aa7ab05ac15fa7ccf53bbc0db